+++ model/Bullet.hpp
@@ -20,11 +20,11 @@ public:
     int playerId;
     Vec2Double position;
     Vec2Double velocity;
-    int damage;
+    double damage;
     double size;
     std::shared_ptr<ExplosionParams> explosionParams;
     Bullet();
-    Bullet(WeaponType weaponType, int unitId, int playerId, Vec2Double position, Vec2Double velocity, int damage, double size, std::shared_ptr<ExplosionParams> explosionParams);
+    Bullet(WeaponType weaponType, int unitId, int playerId, Vec2Double position, Vec2Double velocity, double damage, double size, std::shared_ptr<ExplosionParams> explosionParams);
     static Bullet readFrom(InputStream& stream);
     void writeTo(OutputStream& stream) const;
     std::string toString() const;
